Simple runscript prototype (#67)
* metarunner bugfix

* add filesystem methods

* update datastructures

* mild refactor of export_glue_data

* moving some files

* Simple script

+++ jiant/proj/main/runscript.py
@@ -11,6 +11,7 @@ import jiant.shared.initialization as initialization
 import jiant.shared.distributed as distributed
 import jiant.shared.model_setup as model_setup
 import jiant.utils.torch_utils as torch_utils
+import jiant.utils.python.io as py_io
 import jiant.utils.zconf as zconf
 
 
@@ -203,14 +204,41 @@ def run_loop(args: RunConfiguration, checkpoint=None):
                 path=os.path.join(args.output_dir, "test_preds.p"),
             )
 
-    if args.delete_checkpoint_if_done and args.save_checkpoint_every_steps:
+    if (
+        args.delete_checkpoint_if_done
+        and args.save_checkpoint_every_steps
+        and os.path.exists(os.path.join(args.output_dir, "checkpoint.p"))
+    ):
         os.remove(os.path.join(args.output_dir, "checkpoint.p"))
+
+    py_io.write_file("DONE", os.path.join(args.output_dir, "done_file"))
 
 
 def run_resume(args: ResumeConfiguration):
-    checkpoint = torch.load(args.checkpoint_path)
+    resume(checkpoint_path=args.checkpoint_path)
+
+
+def resume(checkpoint_path):
+    checkpoint = torch.load(checkpoint_path)
     args = RunConfiguration.from_dict(checkpoint["metadata"]["args"])
     run_loop(args=args, checkpoint=checkpoint)
+
+
+def run_with_continue(cl_args):
+    run_args = RunConfiguration.default_run_cli(cl_args=cl_args)
+    if os.path.exists(os.path.join(run_args.output_dir, "done_file")) or os.path.exists(
+        os.path.join(run_args.output_dir, "val_metrics.json")
+    ):
+        print("Already Done")
+        return
+    elif run_args.save_checkpoint_every_steps and os.path.exists(
+        os.path.join(run_args.output_dir, "checkpoint.p")
+    ):
+        print("Resuming")
+        resume(os.path.join(run_args.output_dir, "checkpoint.p"))
+    else:
+        print("Running from start")
+        run_loop(args=run_args)
 
 
 def main():
@@ -219,6 +247,8 @@ def main():
         run_loop(RunConfiguration.default_run_cli(cl_args=cl_args))
     elif mode == "continue":
         run_resume(ResumeConfiguration.default_run_cli(cl_args=cl_args))
+    elif mode == "run_with_continue":
+        run_with_continue(cl_args=cl_args)
     else:
         raise zconf.ModeLookupError(mode)
